# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Money Graph API",
    "description": "One completed run. Synthetic data requires explicit fixture mode; see X-Data-Source.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/nodes/{gid}": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "Get Node",
        "operationId": "get_node_api_nodes__gid__get",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Gid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Identifier or export not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Stale AI run_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No completed run available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/priorities": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "Get Priorities",
        "operationId": "get_priorities_api_priorities_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrioritiesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Identifier or export not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Stale AI run_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "No completed run available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error",
          "run_id"
        ],
        "title": "ErrorResponse"
      },
      "Evidence": {
        "properties": {
          "rule_id": {
            "type": "string",
            "minLength": 1,
            "title": "Rule Id"
          },
          "metric": {
            "type": "string",
            "minLength": 1,
            "title": "Metric"
          },
          "operator": {
            "type": "string",
            "enum": [
              "gt",
              "gte",
              "lt",
              "lte",
              "eq"
            ],
            "title": "Operator"
          },
          "actual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Threshold"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule_id",
          "metric",
          "operator",
          "actual",
          "threshold",
          "passed"
        ],
        "title": "Evidence"
      },
...
      "NodeMetrics": {
        "properties": {
          "in_degree_unique": {
            "type": "integer",
            "minimum": 0.0,
            "title": "In Degree Unique"
          },
          "out_degree_unique": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Out Degree Unique"
          },
          "tx_in_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tx In Count"
          },
          "tx_out_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tx Out Count"
          },
          "in_amount_kzt": {
            "type": "string",
            "pattern": "^[0-9]+(?:\\.[0-9]+)?$",
            "title": "In Amount Kzt"
          },
          "out_amount_kzt": {
            "type": "string",
            "pattern": "^[0-9]+(?:\\.[0-9]+)?$",
            "title": "Out Amount Kzt"
          },
          "out_in_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Out In Ratio"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "in_degree_unique",
          "out_degree_unique",
          "tx_in_count",
          "tx_out_count",
          "in_amount_kzt",
          "out_amount_kzt",
          "out_in_ratio"
        ],
        "title": "NodeMetrics"
      },
      "NodeQuality": {
        "properties": {
          "is_seed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Seed"
          },
          "hop_depth": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Hop Depth"
          },
          "outbound_censored": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outbound Censored"
          },
          "inbound_incomplete": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inbound Incomplete"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "is_seed",
          "hop_depth",
          "outbound_censored",
          "inbound_incomplete",
          "reasons"
        ],
        "title": "NodeQuality"
      },
      "NodeRecord": {
        "properties": {
          "gid": {
            "type": "string",
            "minLength": 1,
            "title": "Gid"
          },
          "component_id": {
            "type": "string",
            "minLength": 1,
            "title": "Component Id"
          },
          "cluster_id": {
            "type": "string",
            "minLength": 1,
            "title": "Cluster Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "consolidator",
              "transit",
              "distributor",
              "terminal",
              "coordinator",
              "peripheral"
            ],
            "title": "Role"
          },
          "role_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Role Score"
          },
          "priority_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority Score"
          },
          "assignment_status": {
            "type": "string",
            "enum": [
              "rule_matched",
              "insufficient_evidence"
            ],
            "title": "Assignment Status"
          },
          "metrics": {
            "$ref": "#/components/schemas/NodeMetrics"
          },
          "quality": {
            "$ref": "#/components/schemas/NodeQuality"
          },
          "role_evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Role Evidence"
          },
          "priority_evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Priority Evidence"
          },
          "role_explanation": {
            "type": "string",
            "title": "Role Explanation"
          },
          "priority_explanation": {
            "type": "string",
            "title": "Priority Explanation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gid",
          "component_id",
          "cluster_id",
          "role",
          "role_score",
          "priority_score",
          "assignment_status",
          "metrics",
          "quality",
          "role_evidence",
          "priority_evidence",
          "role_explanation",
          "priority_explanation"
        ],
        "title": "NodeRecord"
      },
      "NodeResponse": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "1.0",
            "title": "Contract Version",
            "default": "1.0"
          },
          "run_id": {
            "type": "string",
            "minLength": 1,
            "title": "Run Id"
          },
          "node": {
            "$ref": "#/components/schemas/NodeRecord"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contract_version",
          "run_id",
          "node"
        ],
        "title": "NodeResponse"
      },
      "PrioritiesResponse": {
        "properties": {
          "contract_version": {
            "type": "string",
            "const": "1.0",
            "title": "Contract Version",
            "default": "1.0"
          },
          "run_id": {
            "type": "string",
            "minLength": 1,
            "title": "Run Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NodeRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contract_version",
          "run_id",
          "items",
          "total"
        ],
        "title": "PrioritiesResponse"
      },
...
    }
  }
}